TRANSMITTED FIRST PPACKETclearclear, no way the problem was my uint64 typedef... oh well LFG it finally works

+++ include/structures.h
@@ -12,7 +12,7 @@ typedef unsigned int size_t;
 typedef unsigned char uint8_t;
 typedef unsigned short uint16_t;
 typedef unsigned int uint32_t;
-typedef unsigned long uint64_t;
+typedef unsigned long long uint64_t;
 //size = 4 + 4 + 4 + 4 = 16
 // int + int + sizeof(ptr) + sizeof(ptr) = 16
 struct MemoryBlock{
